refactor: nfc class
